Move CollectionViewBase to a (new) collection_view module.
